+++ pkg/feed/fetching/client.go
@@ -4,15 +4,18 @@
 package fetching
 
 import (
-	"bytes"
 	"context"
+	"encoding/xml"
+	"errors"
+	"fmt"
 	"io"
 	"net/http"
+	"strings"
 	"time"
+	"unicode"
 
 	"github.com/cespare/xxhash/v2"
 	"github.com/mmcdole/gofeed"
-	"github.com/rs/zerolog/log"
 )
 
 // A Client performs outgoing HTTP requests to get remote feed data.
@@ -35,7 +38,7 @@ func NewClient(httpClient *http.Client, userAgent string) *Client {
 	}
 }
 
-// Fetch performs a HTTP request to get feed data and parses it.
+// Fetch performs an HTTP request to get feed data and parses it.
 //
 // Adapted from gofeed.Parser.ParseURL with the following modifications:
 // - User-Agent header;
@@ -44,12 +47,7 @@ func NewClient(httpClient *http.Client, userAgent string) *Client {
 func (c *Client) Fetch(ctx context.Context, feedURL string, eTag string, lastModified time.Time) (FeedStatus, error) {
 	req, err := http.NewRequestWithContext(ctx, http.MethodGet, feedURL, nil)
 	if err != nil {
-		log.
-			Error().
-			Err(err).
-			Str("feed_url", feedURL).
-			Msg("feed: failed to create request")
-		return FeedStatus{}, err
+		return FeedStatus{}, fmt.Errorf("feed: failed to create request: %w", err)
 	}
 
 	req.Header.Set("User-Agent", c.userAgent)
@@ -64,12 +62,7 @@ func (c *Client) Fetch(ctx context.Context, feedURL string, eTag string, lastMod
 
 	resp, err := c.httpClient.Do(req)
 	if err != nil {
-		log.
-			Error().
-			Err(err).
-			Str("feed_url", feedURL).
-			Msg("feed: failed to perform request")
-		return FeedStatus{}, err
+		return FeedStatus{}, fmt.Errorf("feed: failed to perform request: %w", err)
 	}
 
 	defer func() {
@@ -99,21 +92,55 @@ func (c *Client) Fetch(ctx context.Context, feedURL string, eTag string, lastMod
 		}
 	}
 
-	var body bytes.Buffer
-	teeReader := io.TeeReader(resp.Body, &body)
-
-	parsedFeed, err := c.feedParser.Parse(teeReader)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
-		log.
-			Error().
-			Err(err).
-			Str("feed_url", feedURL).
-			Msg("feed: failed to parse feed")
+		return FeedStatus{}, fmt.Errorf("feed: failed to read response body: %w", err)
+	}
+
+	parsedFeed, err := c.parse(body)
+	if err != nil {
 		return FeedStatus{}, err
 	}
 
 	feedStatus.Feed = parsedFeed
-	feedStatus.Hash = xxhash.Sum64(body.Bytes())
+	feedStatus.Hash = xxhash.Sum64(body)
 
 	return feedStatus, nil
+}
+
+// parse wraps gofeed.Parser to handle minor parsing errors for feeds containing improperly formatted data
+// or invalid Unicode characters.
+func (c *Client) parse(body []byte) (*gofeed.Feed, error) {
+	feedStr := string(body)
+	parsedFeed, err := c.feedParser.ParseString(feedStr)
+
+	var xmlSyntaxError *xml.SyntaxError
+	if errors.As(err, &xmlSyntaxError) {
+		// Atom and RSS feed parsing errors.
+
+		if strings.HasPrefix(xmlSyntaxError.Msg, "illegal character code") {
+			// The feed contains non-printable Unicode characters.
+			//
+			// Filter the input to remove non-printable Unicode characters, then attempt to parse feed data.
+			//
+			// See:
+			// - https://pkg.go.dev/unicode#IsPrint
+			// - https://pkg.go.dev/unicode#IsGraphic
+			// - https://pkg.go.dev/unicode#RangeTable
+			filtered := strings.Map(func(r rune) rune {
+				if unicode.IsPrint(r) {
+					return r
+				}
+				return -1
+			}, feedStr)
+
+			parsedFeed, err = c.feedParser.ParseString(filtered)
+		}
+	}
+
+	if err != nil {
+		return &gofeed.Feed{}, err
+	}
+
+	return parsedFeed, nil
 }